requirements.txt
pandas
numpy
yfinance
alpaca-py